refactor(proto): split agent config tests
