+++ nodelite-proto/src/config/tests.rs
@@ -11,9 +11,10 @@ use super::{
     DEFAULT_AUDIT_RETENTION_DAYS, DEFAULT_GEOIP_UPDATE_INTERVAL_DAYS, DEFAULT_MAX_MESSAGE_BYTES,
     DEFAULT_WS_AUTH_BLOCK_SECS, DEFAULT_WS_AUTH_FAIL_MAX_ATTEMPTS,
     DEFAULT_WS_AUTH_FAIL_WINDOW_SECS, DEFAULT_WS_MAX_CONNECTIONS_PER_IP,
-    DEFAULT_WS_MAX_TOTAL_CONNECTIONS, GeoIpEdition, GeoIpProvider, MAX_NODE_IDENTITY_TEXT_BYTES,
-    MAX_NODE_TAG_BYTES, parse_agent_config, parse_server_config,
+    DEFAULT_WS_MAX_TOTAL_CONNECTIONS, GeoIpEdition, GeoIpProvider, parse_server_config,
 };
+
+mod agent;
 
 #[test]
 fn server_example_documents_install_section() {
@@ -302,107 +303,6 @@ fn rejects_invalid_trusted_proxy_cidr() {
     .expect_err("invalid trusted proxy cidr should fail");
 
     assert!(error.to_string().contains("server.trusted_proxies"));
-}
-
-#[test]
-fn rejects_invalid_agent_server_scheme() {
-    let error = parse_agent_config(
-        r#"
-        [agent]
-        node_id = "hk-01"
-        node_label = "Hong Kong 01"
-        server = "http://127.0.0.1:8080/ws"
-        token = "token"
-        "#,
-    )
-    .expect_err("invalid agent config should fail");
-
-    assert!(error.to_string().contains("agent.server"));
-}
-
-#[test]
-fn parses_agent_config() {
-    let config = parse_agent_config(
-        r#"
-        [agent]
-        node_id = "hk-01"
-        node_label = "Hong Kong 01"
-        server = "ws://127.0.0.1:8080/ws"
-        token = "token"
-        report_interval_secs = 7
-        hostname_override = "hk-01.internal"
-        tags = [" edge ", "apac"]
-        "#,
-    )
-    .expect("agent config should parse");
-
-    assert_eq!(config.node_id, "hk-01");
-    assert_eq!(config.report_interval_secs, 7);
-    assert_eq!(config.tags, vec!["apac", "edge"]);
-}
-
-#[test]
-fn rejects_agent_config_with_too_many_tags() {
-    let tags = (0..1000)
-        .map(|index| format!("\"tag-{index}\""))
-        .collect::<Vec<_>>()
-        .join(", ");
-    let input = format!(
-        r#"
-        [agent]
-        node_id = "hk-01"
-        node_label = "Hong Kong 01"
-        server = "ws://127.0.0.1:8080/ws"
-        token = "token"
-        tags = [{tags}]
-        "#
-    );
-
-    let error = parse_agent_config(&input).expect_err("too many tags should fail");
-    assert!(error.to_string().contains("agent.tags"));
-}
-
-#[test]
-fn rejects_agent_config_with_oversized_tag() {
-    let oversized = "x".repeat(MAX_NODE_TAG_BYTES + 1);
-    let input = format!(
-        r#"
-        [agent]
-        node_id = "hk-01"
-        node_label = "Hong Kong 01"
-        server = "ws://127.0.0.1:8080/ws"
-        token = "token"
-        tags = ["{oversized}"]
-        "#
-    );
-
-    let error = parse_agent_config(&input).expect_err("oversized tag should fail");
-    assert!(error.to_string().contains("agent.tags[0]"));
-}
-
-#[test]
-fn rejects_agent_config_with_oversized_identity_text() {
-    let oversized = "界".repeat(MAX_NODE_IDENTITY_TEXT_BYTES / 3 + 1);
-    for (field, extra) in [
-        ("agent.node_label", format!("node_label = \"{oversized}\"")),
-        (
-            "agent.hostname_override",
-            format!("node_label = \"Hong Kong 01\"\nhostname_override = \"{oversized}\""),
-        ),
-    ] {
-        let input = format!(
-            r#"
-            [agent]
-            node_id = "hk-01"
-            {extra}
-            server = "ws://127.0.0.1:8080/ws"
-            token = "token"
-            "#
-        );
-
-        let error = parse_agent_config(&input).expect_err("oversized identity text should fail");
-        assert!(error.to_string().contains(field));
-    }
 }
 
 #[test]
